requirements.txt
pandas==1.2.0
numpy==1.19.2
matplotlib==3.3.3
Pillow==8.1.0
plotly==4.14.3
scikit_learn==0.24.1
seaborn==0.11.1
streamlit==0.75.0
